v2.0

+++ src/shared/infra/database/migrations/1631032236330-CreateAddTrades.ts
@@ -36,7 +36,7 @@ export class CreateAddTrades1631032236330 implements MigrationInterface {
           },
           {
             name: 'value_trade',
-            type: 'integer',
+            type: 'float',
           },
           {
             name: 'created_at',
